Create initial version of Content Protection application
This is a simple smart contract written for Hyperledger Sawtooth.
It is only to understand and demonstrate the concept and not for production.
Note: There are many TODOs, important being the way to change READ handler.

Content Protection application provides User (registration, update), Content
(creation, updation, read) based on roles.

Signed-off-by: S m, Aruna <aruna.s.m@intel.com>